% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{cellboard Shiny server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Standard Shiny server arguments.}
}
\value{
Used for its side effects.
}
\description{
cellboard Shiny server
}
